Add support for ropevim_goto_def_newwin variable
As per doc/readme.txt there should be an option to define where should
be a new found definition opened. Seems like the implementation is
missing in current tree. In ropemode the variable is passed via 'other'
parameters. This patch uses 'other' in case it fits into words defined
in readme.txt ('new' or 'vnew') otherwise it opens in current buffer.

Resolves: https://github.com/python-rope/ropevim/issues/11